requirements.txt
pytest == 7.2.1
allure-pytest == 2.12.0
PyYAML == 6.0
requests == 2.28.1